make testing-utils independent of ledger-api
move non ledger-api stuff to libs-scala

remove some dead code + move classes to simplify deps
